refactor(plan): split planaggregate event records into separate files
Move PlanGenerated, MesoCycleCreated, and FirstMicroCycleCreated each
into their own file to satisfy the one-type-per-file convention.
Remove the SA1402/SA1649 pragma suppressions from PlanAggregate.cs.
